MP2-3980 fix current PR finance values stored with wrong ids

+++ jems-server/src/main/kotlin/io/cloudflight/jems/server/project/service/report/project/financialOverview/ProjectReportCertificateLumpSumPersistence.kt
@@ -13,5 +13,5 @@ interface ProjectReportCertificateLumpSumPersistence {
 
     fun updateCurrentlyReportedValues(projectId: Long, reportId: Long, currentValues: Map<Int, BigDecimal>)
 
-    fun updateCurrentlyVerifiedValues(projectId: Long, reportId: Long, verifiedValues: Map<Long, BigDecimal>)
+    fun updateCurrentlyVerifiedValues(projectId: Long, reportId: Long, verifiedValues: Map<Int, BigDecimal>)
 }
